feat: adapted trainer and predictor to manage batch inference and return evaluation results in vertex ai schema
