+++ trainer/src/trainer/config.py
@@ -1,3 +1,5 @@
+"""This module defines the configuration for the training job."""
+
 import os
 from typing import Any, Literal
